requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --no-annotate
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.3
mdurl==0.1.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pygments==2.19.1
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
rich==14.0.0
six==1.17.0
structlog==25.3.0